Prepare for moving user data/config to home directories (patch #2949378)
Patch description: These two patches together form a proposed first step for
moving user data from it's current location -- scattered inside the application
base directory -- and out to locations under the users home. This is better for
many reasons, some of which are: the upgrade process is easier, the current
scheme can't be easily packaged by (for example) Debian and placing stuff in the
users home just makes more sense.

Previously many paths were hard coded into:

bookkeeping.data.util.SSFileSystem

The location of this class is a bit confusing and counter-intuitive. Most of the
rest of the package bookkeeping.data concerns user data; it's silly to put paths
to GUI graphics and icons next to that.

In these patches a new package bookkeeping.app is introduced. I think it makes
more sense to have this stuff there. I am in the process of re-factoring some of
the configuration code, and this is a convenient place to put it: long away from
the database details. Also, it doesn't really fit in anywhere else (paths are
not an util).

The first patch creates the necessary changes, introduces the SSPath class
(with tests), app hierarchy and adds some code from the freedesktop project
to manage the locations. (The code is yanked from
http://people.freedesktop.org/~krake/xdg-java/)

The second patch removes all references to the SSFileSystem, thus making it safe
to completely remove that class. Also it cleans up unecessary usage of String
where it's in fact a File object we want.

The changes in behavior is kept minimal: (Nothing is changed on MacOS X or
Windows)

* The first patch introduces code to automatically create the necessary
directories on GNU/Linux and other Unix-like OS:es. (That would be the
freedesktop.org compliant ~/.local/share/jfsaccounting and
~/.config/jfsaccounting) Also it @deprecates SSFileSystem.

* The second patch moves the backup.history file to the
.local/share/jfsaccounting directory. This was used during testing and is
working just fine. All other configuration and data files remain in place.

git-svn-id: svn://svn.code.sf.net/p/jfsaccounting/code/trunk@64 bfc8530a-dd38-4376-9a49-ec652ef1b1b4